+++ odoo/local-src/specific_project/controllers/__init__.py
@@ -1,3 +1,4 @@
 from . import reference
 from . import reference_form
 from . import reference_search
+from . import helpers