+++ HISTORY.md
@@ -9,6 +9,8 @@
 * [RM#16738](https://redmine.iart.ch/issues/16738) publish imp for partner
     + fix parter/user relations
 * Update `cms_delete_content`: delete confirmation now happens in modal
+* [RM#16412](https://redmine.iart.ch/issues/16412) references aggregation: random
+  Includes partial refactoring of mosaic JS.
 
 
 ## 9.2.4b4 (2017-03-08 - INT)